@@ -7,53 +7,34 @@ import film.api.models.User;
 import film.api.service.ChapterService;
 import film.api.service.HistoryService;
 import film.api.service.UserService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.util.StringUtils;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
-import film.api.models.Role;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import javax.annotation.security.RolesAllowed;
-import javax.servlet.http.HttpServletRequest;
-import java.awt.*;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.StandardCopyOption;
-import java.nio.file.*;
+import jakarta.servlet.http.HttpServletRequest;
 
 @RestController
+@RequiredArgsConstructor
 public class UserController {
 
     @Value("${jwt.header}")
     private String tokenHeader;
-
-    @Autowired
-    private JWTUtil jwtUtil;
-
-    @Autowired
+    private final JWTUtil jwtUtil;
     @Qualifier("jwtUserDetailsService")
-    private UserDetailsService userDetailsService;
-    @Autowired
+    private final UserDetailsService userDetailsService;
+    private final JwtUserDetailsService jwtUserDetailsService;
+    private final HistoryService historyService;
+    private final PasswordEncoder passwordEncoder;
+    private final UserService userService;
 
-    private JwtUserDetailsService userDetailsServicea;
-    @Autowired
-    private HistoryService historyService;
-    @Autowired
-    private ChapterService chapterService;
-    @Autowired
-    private UserService userService;
     @RequestMapping(value = "user", method = RequestMethod.GET)
-
     @Secured({"ROLE_ADMIN", "ROLE_USER"})
     public JwtUser getAuthenticatedUser(HttpServletRequest request) {
         String token = request.getHeader(tokenHeader).substring(7);
@@ -63,7 +44,7 @@ public class UserController {
 
     @RequestMapping(value = "/register", method = RequestMethod.POST)
     public ResponseEntity<?> saveUser(@RequestBody UserSignupDTO user) throws Exception {
-        return new ResponseEntity<>(userDetailsServicea.save(user),HttpStatus.CREATED);
+        return new ResponseEntity<>(jwtUserDetailsService.save(user, passwordEncoder),HttpStatus.CREATED);
     }
     @Secured({"ROLE_ADMIN", "ROLE_USER"})
     @GetMapping("/ApiV1/UserByLogin")
