@@ -1,15 +1,11 @@
 package film;
 
 import film.api.helper.FileSystemHelper;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.http.CacheControl;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -19,12 +15,14 @@ import java.nio.file.Paths;
 @Configuration
 public class BootifulApplication {
 
+	private static final Logger logger = LoggerFactory.getLogger(BootifulApplication.class);
+
 	public static void main(String[] args) {
 		try {
 			setup();
 			SpringApplication.run(BootifulApplication.class, args);
 		} catch (Exception e) {
-			System.out.println("error setting up service, shutting down ...");
+			logger.error("An error occurred while starting the application", e);
 		}
 	}
 
