@@ -1,66 +1,23 @@
 package film.api.service;
 
 import film.api.DTO.ActorDTO;
-import film.api.DTO.UserByAdminDTO;
 import film.api.models.Actor;
-import film.api.models.Chapter;
-import film.api.models.User;
-import film.api.repository.ActorRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-@Service
-public class ActorService {
-    @Autowired
-    private ActorRepository actorRepository;
-    public List<Actor> getList(){
+public interface ActorService {
+    List<Actor> getList();
 
-        return actorRepository.findAll();
-    }
+    Actor addActor(Actor actor);
 
-    public Actor addActor(Actor actor) {
-        return actorRepository.save(actor);
-    }
+    Optional<Actor> findById(Long id);
 
-    public Optional<Actor> findById(Long id) {
-        return actorRepository.findById(id);
-    }
+    Actor updateActor(Long id, ActorDTO actorPatchDTO);
 
+    void deleteById(Long id);
 
-    public Actor updateActor(Long id, ActorDTO actorPatchDTO) {
+    List<Actor> searchActors(String actorName);
 
-        Actor actor = actorRepository.findById(id).orElse(null);
-
-        if(actorPatchDTO.getSex() != null) {
-            actor.setSex(actorPatchDTO.getSex());
-        }
-        if(actorPatchDTO.getNativeLand() != null) {
-            actor.setNativeLand(actorPatchDTO.getNativeLand());
-        }
-        if(actorPatchDTO.getAge() != null) {
-            actor.setAge(actorPatchDTO.getAge());
-        }
-        if(actorPatchDTO.getActorName() != null) {
-            actor.setActorName(actorPatchDTO.getActorName());
-        }
-
-        return actorRepository.save(actor);
-    }
-
-
-    public void deleteById(Long id) {
-        actorRepository.deleteById(id);
-    }
-
-    public List<Actor> searchActors(String actorName) {
-        return actorRepository.searchActors(actorName);
-    }
-
-    public List<Actor> findActorByChapterId(Long chapterId){
-        return actorRepository.findActorByChapterId(chapterId);
-    }
+    List<Actor> findActorByChapterId(Long chapterId);
 }
